requirements.txt
albumentations==1.3.1
huggingface_hub==0.15.1
matplotlib==3.6.1
numpy==1.23.0
opencv_python_headless==4.7.0.72
pandas==1.5.2
Pillow==9.3.0
pytorch_lightning==1.6.5
scikit_learn==1.2.2
torch==1.13.0
torchvision==0.14.0
ultralytics
transformers